separated quiz utils
