+++ app/page.tsx
@@ -10,8 +10,7 @@ import Link from "next/link";
 import PrimaryButton from "@/components/Button";
 import Loading from "./loading";
 import { formatCurrentTime, formatSessionTime, formatQuizSessionTime } from "@/utils/dateUtils";
-import { getDocs, collection, query, where } from "firebase/firestore";
-import { db } from "@/app/firebaseConfig";
+import { generateQuizLinks } from "@/utils/quizUtils";
 
 export default function Home() {
   const { loggedIn, userId, userDbId } = useAuth();
@@ -70,53 +69,6 @@ export default function Home() {
       setLiveClasses(liveClassesData);
     } catch (error) {
       console.error("Error fetching user sessions:", error);
-    }
-  };
-
-  const generateQuizLinks = async (batchData: any): Promise<Quiz[]> => {
-    try {
-      const sessionsCollection = collection(db, "Sessions");
-
-      const quizObjectsArray = await Promise.all(batchData.map(async (batchItem: any) => {
-        const batchId = batchItem.batch_id;
-
-        const batchQuery = query(sessionsCollection, where("batch", "==", batchId));
-        const querySnapshot = await getDocs(batchQuery);
-        const sessionsDataArray = querySnapshot.docs.map((doc) => doc.data());
-
-        const today = new Date().toISOString().split("T")[0];
-
-        const quizObject = sessionsDataArray
-          .filter((sessionData: any) => sessionData.startDate === today)
-          .map((sessionData: any) => {
-            const redirectParams = sessionData.redirectPlatformParams;
-            if (redirectParams && redirectParams.id) {
-              return {
-                batch: sessionData.batch,
-                end_date: sessionData.endDate,
-                end_time: sessionData.endTime,
-                redirectPlatformParams: {
-                  id: redirectParams.id,
-                },
-                start_date: sessionData.startDate,
-                start_time: sessionData.startTime,
-                redirectPlatform: sessionData.redirectPlatform,
-                name: sessionData.name,
-              };
-            }
-            return null;
-          })
-          .filter((quizObject) => quizObject !== null) as Quiz[];
-
-        return quizObject;
-      }));
-
-      const flattenedQuizObjectsArray = quizObjectsArray.flat();
-
-      return flattenedQuizObjectsArray;
-    } catch (error) {
-      console.error("Error fetching quiz data from Firestore:", error);
-      throw error;
     }
   };
